update signIn and remove Login
